refactor(claim-gates): split scanners into license/numeric/prose/fleet modules
Second C12 length pass: break claim-drift/scanners.ts (~2.3k) into focused
modules with a thin barrel. No public API change.
